% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mta.R
\docType{data}
\name{mta}
\alias{mta}
\title{MTA Tweets}
\format{An object of class \code{data.frame} with 90 rows and 50 columns.}
\usage{
mta
}
\description{
ThanksgivingdayTwitter:::mta
}
\details{
MTA's tweets on timeline and replies to riders
}
\keyword{datasets}
