% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt.R
\docType{data}
\name{rt}
\alias{rt}
\title{Riders' tweets about MTA during 24hrs on Thanksgiving day}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 10092 rows and 42 columns.}
\usage{
rt
}
\description{
ThanksgivingdayTwitter:::rt
}
\details{
Parsed dataset from test-stream.json file
Dataset includes riders' tweets about MTA
}
\keyword{datasets}
